Add IsChineseOrPunction api
